requirements.txt
pytest==8.3.3
numpy==1.26.4
pillow==10.4.0
kaggle==1.6.17
scikit-learn==1.5.2
matplotlib==3.8.4
mlflow==2.12.1
fastapi==0.115.5
uvicorn==0.30.6
prometheus-client==0.20.0
python-multipart==0.0.9
